+++ CHANGELOG.md
@@ -12,6 +12,7 @@ For a full diff see [`1.0.0...main`][1.0.0...main].
 
 - Added `methodsAllowedToUseContainerTypeDeclarations` parameter to allow configuring a list of method names that are allowed to have container parameter type declarations ([#541), by [@localheinz]
 - Allowed disabling rules ([#542), by [@localheinz]
+- Added support for nullable union types ([#543), by [@localheinz]
 
 ### Changed
 
@@ -470,6 +471,7 @@ For a full diff see [`362c7ea...0.1.0`][362c7ea...0.1.0].
 [#540]: https://github.com/ergebnis/phpstan-rules/pull/540
 [#541]: https://github.com/ergebnis/phpstan-rules/pull/541
 [#542]: https://github.com/ergebnis/phpstan-rules/pull/542
+[#543]: https://github.com/ergebnis/phpstan-rules/pull/543
 
 [@enumag]: https://github.com/enumag
 [@ergebnis]: https://github.com/ergebnis